requirements.txt
pyrogram==2.0.106
tgcrypto<=1.2.3
aiohttp<=3.8.1
python-dotenv<=0.20.0
motor
aiofiles
dnspython
apscheduler
requests
psutil
pyromod
